The HSD regression reported for this refactor did not get fixed by the backout. The refactor is unlikely to have caused the TDR that was reported. Hence, rechecking in the change.
The CL 654459 was a refactor for fixing a TIL issue. In that CL, I changed the way IGC receives and stores AIL info from UMD.

With this CL, I revert the changes made in that CL and instead store the use the m_uscAilInfo variable from the translationContext itself.

This will isolate the DX specific var from the platforms file.

Change-Id: Id98d7113cc04ab336ddbe75505674283f6133ecd