Change signatures of functions returning structures or taking structures as an argument.
When Clang compiles functions that take or return structures, it replaces the signature of the function to work on pointers. The same was done on SPIR-V path, but only for built-in functions. This change introduces this behaviour also for functions defined in current module. This is needed in cases where we would like to link modules genereted by Clang with modules translated with SPIR-V reader, e.g. for some extension functions.

Change-Id: I063fc88aa5eaab4a2f30d846382f982b3d38bb32