Prevent loop spliting in LLVM LoopSimplification pass. When barriers are present if LLVM pass decides to split the barrier we may end up with divergent barriers causing functional issues..
In term of performance it can also cause low SIMD utilization.
Add a pass to canonicalize the loop in our prefered way before
running loop passes.

LLVM shouldn't break loop with convergent instructions, the
Simplification pass should get fixed to prevent this kind of
issue. I'll try to prepare a patch to see if this can be
prevented.

Change-Id: I2881aab5e85ea53763b5b788260fc4c523c2055c