Changing VME types to opaque.
In new Clang versions VME types are built-in. They will be declared as opaque structure types. This change prepares IGC for this change without breaking compatibility with Clang based on LLVM4.
This is achieved by adding conversion helpers that are resolved on LLVM level.

In CShader.cpp optimisation for GRF size of VME payload is reverted, as it was not fully functionally correct. This will be addressed in future commits

Change-Id: I5f085adf9e9732da5a3511c98f9f23e25ba7e528