Moving the localID present information to the new metadata framework..
With this change, we add the function pointer to the new
framework. whenever we see localID present and set the bool
localIDPresent = true. for that FunctionMetaData object..

Later, when we want to read the localIDPresent, we check if we
had. populated the new framework with that function pointer or
not.. If we had not, then that means local ID present was never
set.. IF we had, then we check that localIDPresent was set or
not.

Change-Id: I2b524d264c073bd17ea698894485cf8101dbcd42