Reason of the issue was that in some cases (when using Shader Debugger that captures all intermediate computations steps to the buffers) we may need to legalize uses of i1 vectors in IGC.. While during Legalization pass IGC legalizes some of the i1 occurences, it doesn't handle vectors correctly, also, this is too late, since earlier pass VectorPreProcess cannot handle i1 vectors correctly.. Therefore, this change consists of two things:.
Moving VectorPreProcess after LegalizationPass - doing the
opposite thing, so moving LegalizationPass before
VectorPreProcess was not an option, because Legalization cleans
up some of the constructs that are introduced later in the
compilation process..

Adding support for i1 vector legalization for stores. This was
added following existing Legalization schema, so when handling
instruction sequence that leads to problematic store. Sequence
starts with insertelement instruction, that also has to be
legalized to be handled by EmitVISAPass. This change is
following the scheme of handling ExtractElement, cleaning up
some of the code along the way.

Change-Id: Iaf01cf8dfd7fda5a570f860b97548530b78cc969