With this change, I skip compiling SIMD modes which are not profitable.
The previous change was to ensure that only the functionally correct
SIMD modes get compiled and returned.

The order of SIMD modes returned is descending order (32, 16,8)
So, there is no default SIMD mode anymore. Hence I have removed it.

The order of checking the conditions is this:
1. Forced SIMD
2. SIMD conditions pass
3. SIMD conditions fail - functional failure
4. SIMD conditions fail - profitability failure

For profitability case, i.e., case 4 above, we need to check whether this is the first functionally
correct SIMD mode or not. If it is the first functionally correct SIMD mode, then we still compile the
SIMD mode for it. This is done by checking the m_canAbortOnSpill. For
first functionally correct SIMD mode, m_canAbortOnSpill will be false,
and we will compile it

Change-Id: Id0195ea4d4ad5eab99d885f4d5a88c97192c0f9f