All nodes of congruent class (CC) are linked. The previous implementation keeps an isolated node in that link. This change will remove an isolated node from the link and put it in a separate congruent class (single-valued, thus isolated).
This change will make it be able to do re-union nodes from different
congruent class later after DeSSA is done. And this change seems
easier to understand as well.

The first version of this change had a bug (did not reset leader
always once a node is removed), which is fixed in this change.
Also, for the existing Path Halving (getLeader), it did not set
the parent node correctly, though this is not correctness issue
but bad for performance, which is also fixed in this change

Change-Id: I8976dd394d24034cef388f4fbe4e2f77be62af79