Stop guessing opencl-clang library name in in-tree build
We hardcoded the name of opencl-clang as:
  libopencl-clang.so.${LLVM_VERSION_MAJOR}
which might not be correct on some LLVM versions. For example, in case
of llvm-7.1.0 we will have:
  libopencl-clang.so.7.1
which will cause IGC build failure in case of in-tree build path. To avoid
that we instead can just reuse existing opencl-clang cmake target instead
of redefining it

Change-Id: I54cd1354abafb77a233b070dceaf35eb9986ecd6