In createImplicitArgs() and non-kernel functions for subroutine/function call handling, function argument is assumed to be root value of dessa (mostly), which is not guaranteed in our implementation. It seems no reason for requiring that either. This change will use GetSymbol() to get CVariable, in which dessa is handled correctly.
The issue was caught in a test case in which a kernel arg is
in a dessa CC (three values) and the arg isn't a root value
(in the old implementation, a root value is just a label
and is not considered as a member of CC. The new implementation
does make the root value a member). The arg was created
a separate cvariable from the rest of dessa CC, which isn't
right in terms of dessa design. Surprisingly, it does not
cause any functional issue until the new phiSrcMov uses
rootValues to check if they are in the same dessa CC. The
check thinks no mov is needed as they are in the same CC,
but the mov is needed because they are actually not using
the same cvariable due to this bug.

This change adds the new functionality, but keeps if off
(no functional change). It will be turned on in the next
commit

Change-Id: Ifdf3e6693cfeb74e1b0aeacf60ea2f4f84c8efee