Remove WaveIntrinsicWAPass
Since SimplyCFG's new SinkCommonInsts option is set to false by default when
createCFGSimplificationPass is called, wave intrinsics in an if/else will
no longer be sunk to their common successor. This removes the need for the
WaveIntrinsicWAPass. The other functions associated with that pass have
been moved to the files that were calling them

Change-Id: Iac4f424248d9fcd2eed7db0f6d693a9b4775ee8e