GenISA_WaveClustered - new intrisic for clustered reduction.
Performs the reduction (accumulation) not on the whole wave,
but on clusters, that is, on a set of sequences of consecutive lanes,
and broadcasts the result to associated output clusters.
For details refer the description provided in the intrinsic definitions file.
This is a generalization of GenISA_WaveAll concept, which can be
considered as GenISA_WaveClustered with cluster size equal to
the size of the whole wave.

Change-Id: Ib6ebd504fd020de2c6fa7a17cd914e8599f935ad