This change is needed from LLVM9 onwards. The issue is related to SCEV causing an increasing stack size over 1MB for certain use cases because of SCEV's recursive structure.
We determine the loop unroll cutoff point by catching a Binary Op -> PHI -> Loop latch pattern
and how many integer ops need to be processed to represent this Binary Op. Once we know this number
we use the Trip Count info to determine unroll potential and calculate an appopriate Threshold.

Change-Id: I20605b8c5b4d70181dbda0f4c74fb0a4e60dc657