[Autobackout]Revert of change: 9fe158e12054a97d703f59bfb75607f646430d80 - Enforce immediate value constraint for inline asm. - Fixed a bug where the source of an inline asm call with multiple outputs was not being matched in PatternMatchPass. - Fixed an issue where LLVM promotes a variable to a constant, causing an immediate value to be written to the asm string when it should not be.
__asm__ ("mov (M1, 16) %0(0,0)<1> %1(0,0)<1;1,0>" : "=rw" (s1) : "rw" (s2) );

When s2 is promoted to a constant value, the following vISA instruction is generated. The immediate value still contains the regioning info from the user inlined string.
mov (M1, 16) V39(0,0)<1> 0x3f800000:f(0,0)<1;1,0>

By enforcing the constraint, if we detect an immediate value with the "rw" constraint, we can generate an extra move instruction to resolve this issue.

Change-Id: Ia4229684d20714c0980106ff39db93eb1aa156e0