Fix build with system LLVM: do not hardcode LLVM library path
The build currently fails on some platforms (namely on Arch Linux)
when using system LLVM:

make[2]: *** No rule to make target '/usr/local/lib/libLLVM-10.so',
needed by 'IGC/Release/elf_packager'.  Stop.

That's because the LLVM link library is now being hardcoded with
a '/usr/local' path on the CMake files since commit 6124673, but
not all systems will install the library on this location.
On Arch Linux, the LLVM library is installed on '/usr/lib',
and find_library() will correctly get the library full path
in a system-agnostic manner.

Also on this commit: continue the CMake execution only if the
LVVM library is found, to avoid such build errors on the middle
of the compilation.