Use Target attribute whenever it is possible, rather than target option.
Cannot replace all uses of target option due to ordering issue. Right
now, attributes are set after G4_kernel has been initialized, but Target
might be used earlier before G4_kernel is initialied (target option
is probably similar to -TotalGRFNum, which is used during G4_kernel
initialization, not after G4_kernel has been initialized).

In the future, we might re-do visa APIs/binary foramt/ordering so
that this kind of issues would not exist. But before that, we have
to live with it.

Change-Id: Iea7cc36b19aa7cdaec79d937774c4849c02f61fc