Remove un-needed CastInst for BinaryOperator and FCmpInst
This patch detects if a BinaryOperator or a FCmpInst has both operands
that are CastInst from float to half and the result of the instruction
is used in a single CastInst from half to float. In this situation, we
don't need this instruction to return a half type because it actually
could return a float and then we can also remove the extra CastInst from
the operands and from the result of this instruction.

Change-Id: If042bae04d39cdfd83160c7f4f2e77a98317ec39