When mixing function pass with module pass as dependent analysis passes, the module analysis pass is considered to be "make change", which causes function analysis passes to be re-run.  As result,  all EmitPass's analysis passes run twice. As some of those analysis passes are time consuming and thus increase compiling time quite a lot.
The module pass used is call graph pass. As this feature is for call handling, the info
(leaf function) is collected in GenXCodeGenModule and saved in GenXFunctionGroupAnalysis.
EmitPass can use GenXFunctionGroupAnalysis (already available to EmitPass) to get this info.
With this, call graph pass is no longer needed.

This change, with the previous submitted change (two analysis passes should return false, not true),
should reduce compiling time.