Add InferAddressSpacesPass only if needed.
The pass is used to propagate address spaces information to memory
operations after the transformation in POC pass (move private memory
allocations to SLM). The pass shouldn't be added into optimization
pipeline if POC pass is disabled.

Also the pass relies on TargetTransformInfoPass and could cause a crash
if dependency is not initialized.