For mergeLoad, make sure to sort mergable loads first so that we will merge as large as possible.
Also, resize() was meant to be before the sort, but it
was after the sort.  With this change, sort is moved before
the loop, and the code looks correct and more readable.

Also, skip updating distance as distance is just a heuristic after
merging (both load/store).

Also, insert merged load into MemRefList so that the list is up
to date and can be used later.