Create FunctionCloningThreshold flag to limit the number of times functions can be cloned.
Functions are cloned for each function group they belong to. This can cause a function to be compiled N times for N function groups that all call this function.
When N is large, this will greatly increase the compile time and memory usage.

This flag limits the number of times a function can be cloned, such that if it's exceeded, we convert it to an indirectly called function.
It will only be compiled once, and runtime relocation is used to patch the function address to each function group which it belongs to.
The default value is 0, which means there is no limit.