For the following code:          mov (8|M0)   r80.0<1>:d    0:w          send (1|M0)  r80:d    r4:uq   0xC            0x041401FF          add (1|M0)   r80.1<1>:d    r54.0<0;1,0>:d    r80.1<0;1,0>:d          add (1|M0)   r80.2<1>:d    r58.0<0;1,0>:d    r80.2<0;1,0>:d          add (1|M0)   r80.3<1>:d    r62.0<0;1,0>:d    r80.3<0;1,0>:d          add (1|M0)   r80.4<1>:d    r66.0<0;1,0>:d    r80.4<0;1,0>:d          add (1|M0)   r80.5<1>:d    r70.0<0;1,0>:d    r80.5<0;1,0>:d          add (1|M0)   r80.6<1>:d    r74.0<0;1,0>:d    r80.6<0;1,0>:d          add (1|M0)   r80.7<1>:d    r78.0<0;1,0>:d    r80.7<0;1,0>:d
Previously, "send" kills r80 entirely, which is incorrect. This change makes
sure that "send (1)" only kills r80.0:d, thus the first "mov" instruction
is not dead.

2nd try: fix an issue that sendRaw's execSize isn't initialized.

This is off by default. Will enable it for testing in the following change.

No functional change expected from this change as it is off.