[Autobackout][FuncReg]Revert of change: 2aa5477d898db8c4e4f1b11ecfe07e07d7d7c9ab * Add BE support for struct arg and return codegen. This is a redesign of the small struct passing optimization for stackcalls.
The previous commit converts struct pointer args into integer data values, however with this method we must use bitwise extraction to get individual struct elements.
The struct arguments, when passed as integers, are stored as AoS format in registers, and we must do a lot of shuffling in the BE to extract all lanes of each element.

This change adds BE support to handle vISA code emission for struct type values. Structs are represented as a vector of BYTES, and are stored as SoA format in registers.
The IGC FE transforms small struct pointers with the "byval" or "sret" attributes back to pass-by-value, and thus selectively undos the SPIRV FE transformation.
Small struct args and retvals are handled directly by vISA codegen in the BE, while larger structs remains pass-by-pointer. Promotable struct sizes can be tuned to minimize spill.