[SPIRVReader] Fix loop metadata generation
In accordance with the SPIR-V spec loop metadata should be attached to
the header block of the loop (via OpLoopMerge or OpLoopControlINTEL)
whilst in LLVM IR it should be attached to latch block of the loop
(via llvm.loop metadata).
Fix attachment place of loop metadata in reverse translation.

This is a port of the following change from Khronos SPIRV-LLVM Translator:
https://github.com/KhronosGroup/SPIRV-LLVM-Translator/commit/abee6f9e8202b7d98b27d2724498aaf431070548