[Autobackout][FuncReg]Revert of change: 590c1bab8bb64ea7393d089baf9b147b0d7d2640  Prepare VCBE to handle standalone stack calls
State before:
The head of a function group is a kernel. Each function that is used in
different function groups (kernels and theirs call graphs) must be
copied. This rule must be applied even for stack calls. In other words,
stack calls are kernel-dependent functions. Such an approach was OK
until there was a need to compile standalone stack calls.

How it is now reorganized:
Each kernel is a head of a function group of type GROUP. Each stack call
(or indirect function) is a head of function group of type SUBGROUP
(legacy, we may consider removal of different types later). Stack calls
won't be copied. Function group passes are run on each function group
independently. This allows the creation of independent stack calls with
no need in a kernel. So, each SUBGROUP is independent function group and
compiles separately, but at the same time, its head may be referenced by
others function groups of type GROUP of SUBGROUP.