Reducing the number of synchronization instructions or their scope
The analysis resulted in eliminating synchronization instructions is based on the position of these instructions in regards of memory access instructions. A synchronization requirements can be omitted in some cases which can be found by deducing if memory stores do not have to be exectued in a particular order or visible for other threads.
SynchronizationObjectCoalescing works in two phases:
- an analysis,
- a transformation.
Such a separation causes to aggregate all transformations in one step which allows to follow all changes in the llvm module.