fix generation of DWARF locations for indirect values
The problem comes from the inconsistency between semantics of <llvm.dbg.value>
and location descriptors defined by DWARF. <llvm.dbg.value> describes the
actual *value*, of a source variable - on the other hand, DWARF location
descriptors focus primarily (with few excepions) on describing the *location*
from which the value can be fetched (the location can be an address
in register file or memory).

Now, DWARF allows for the following location desciptors for the pieces of
variables:
    1. empty location
    2. register location
    3. memory location
    4. implicit location

The situation that was not processed correctly looked like this:

```
call void @llvm.dbg.value(metadata [4096 x i32] addrspace(1)* %0,
                          metadata !1299,
                          metadata !DIExpression(DW_OP_deref))
```

This means that the actual *value* of the source varible !1299 is located
at the address defined by %0. Let's assume %0 is held in a register - then
a proper DWARF location descripor should be of "memory location" type which is
evaluated to the value held in %0. In general case, the respected location
descripor should look like this:

```
    0: DW_OP_reg<N + 16>
    1: DW_OP_const1u <offset in GRF N>
    2: DW_OP_const1u <size in bits of %0>
    3: DW_OP_INTEL_push_bit_piece_stack
```

NOTE_1: Item #3, transforms the original "register location" defined by Item #0
to a "memory location". Such transformations are not possible in the vanilla
DWARF spec, thus we had to opt for intel extensions.

NOTE_2: It seems that we have some level of ambiguity. For example,
I think it is perfectly legal to have this kind of expressions:

```
call void @llvm.dbg.value(metadata i32 %0,
                          metadata !1299,
                          metadata !DIExpression(DW_OP_deref), DW_OP_plus_uconst, 3)
```

Such an expression defines a concrete value. But our current implementaion won't
be able to handle this. From dwarf point of view the respected expression
should be an implicit value. While in reality our backend will emit a
<memory location expression>.
I've added some asserts and diagnostics to detect such cases.