Remove duplication from GenXOCLRuntimeInfo
Symbol and relocation data is only stored in zebin format in VCCodeGen.
Zebin format is translated into legacy one just before the binary
generation. Storing data, symbols and relocations is unified for code,
constant and global variables. Only code relocations are left duplicated
as it is not possible with the current Finalizer iface.