discard *sret* attribute when CMABI transforms a function
CMABI can transform function's return value (for example, when we do
globals localization).  However, LLVM requires functions with a parameter
that has *sret* attribute to have a *void* return type. If this
condition is violated - the IR is considered incorrect.
To preserve IR correctness we discard *sret* attribute when functions
are transformed by CMABI pass.