Generic pointer support breaks address's canonical form
Generic pointer support is implemented by encoding a pointer's
address space into the bits [60:63] of an address. In the current
implementation, when the address is needed, for example for send
instruction, these upper bits are zeroed. But in case when the
actual address has upper bits set (address is allocated from upper
memory), this zeroing breaks the cannonical form requirement. That
could lead to HW fault. The implemented workaround replicates the
bits [56:59] to bits [60:63] assuming the initial address was in
the canonical form.
_OS_DESCRIPTION