Skip add implicit arg metadata when IA=0 to allow for intrinsic usage
When the flag ForceInlineStackCallWithImplArg(IA)=0, we don't need to force inlining, and can use
intrinsics to represent implicit arg usage inside stack-called functions.
When IA=1, we need to write the implicit arg metadata such that we can detect and inline stack calls
that has implicit arguments in its function signature.