Value tracking improvements
Added following improvements to ValueTracker:
1. When previous implementation of findAlloca found memcpy
call, it returned the copied value immediately to tracking loop.
Other current users in findAlloca loop were disregarded this
way and could lead to not finding the tracked value.
2. findAlloca returned nullptr if any of GEP operand was
non-constant. This could be relaxed to returning nullptr in
case of non-constant operand below the current searched
GEP depth.