ValueTracking: do not return nullptr in case non-const index is detected
In case a part of structure that didn't contain image was modified with dynamic
index, we could lose opportunity to track the value.
When we return nullptr, the BTI is set to 0 and we end up with wrong access. It
is better to continue searching, as no high-level API is allowed to access
images dynamically currently.