Don't use relocation for global variables in kernel functions
When zebin is enabled, we would always use relocation for global variables.
This is not necessary for kernels, since we always have access to the global
surface address as an implicit arg, so we can save on relocations in the kernel.
However, when zebin is enabled, we allow relocations in subroutines.