Refactor analysis in CMImplParam pass
Indirect implicit arg usage analysis is refactored. Analysis results
and helper structures are no longer stored in the pass class. Which
makes printing this pass impossible, though this pass is not an analysis
so it won't be printed. On the other hand explicit passing of analysis
results make it more obvious where ones are constructed and where ones
are used. This will make it easier to extend the codebase to support
implicit args in external functions.