[Autobackout][FuncReg]Revert of change: 6c5e956d55701383dd26e2741f0c0cd5114ad362  Analyse callgraph to find max stack call private memory usage
This change does a callgraph analysis to find the max private memory used by a function group.
This helps codegen to determine how much private memory allocation to request from the driver
and decrease the chance of stack overflow.

Note that if there are indirect calls or recursions, we still would not know the call depth.
In this case, we simply add 4KB to the PrivateMemoryPerWI field and hope overflow does not
occur, until better overflow detection is implemented.