VC always emits kernels that support debugging
In order for kernel to be debuggable, the following conditions must be
met:
1. SIP Surface must be allocated
2. The kernel should not use BTI=0

This commit makes sure that we have all these conditions met for non-CM
runtime kernels.

IProgramContext interface is changed in order to decouple
"debuggability" property from the presence of debug information.
VC backend makes use of this change, while Scalar IGC still assumes
that a kernel is debuggable only when debug information is present.