introduce default-overridable options to VC controls
Some control flags should have compiler-defined default value with
an ability to override it. For example, some backend options depend
on enabled optimizatoins.

The commit introduces a concept of "defaultable" controls: if value
is not set (or set to "default"), then compiler defines the value to use
otherwise we take user-provided one,