sginvoke: pass ESIMD text visa assembly in the input of SPMD compilation
Linking of SPMD and ESIMD module is done in CISABuilder when SPMD module is compiled.
This way other functionality related to SPMD compilation like handling
function pointers, ShaderOverride etc don't need to be duplicated in Visa Linker Driver.
Other approach would be to refactor out the common parts but it seems to be very
error-prone at this stage. It can be done when we have more test coverage of
sginvoke functionality.