Add trampoline insertion pass
GenXTrampolineInsertion pass splits a function into two versions if
necessary. One version is for indirect only calls and another is for
direct only calls. Such separation is requested by visa.

If the pass is enabled, it is expected that all calls to functions in
the module that are not present in this module will be indirect.