Add a visa flag -keepScalarJmp
Add a visa flag:  -keepScalarJmp.  Its enum is vISA_KeepScalarJmp.

Default of this flag is false, thus no functional change.  If it is set to
true,  any scalar jmp in visa input will remain as scalar jmp.

This is used for eu fusion platform in which visa converts all scalar jmp into
goto by default.  With this flag, the user has a way to keep all its scalar jmp.