Consider all cases in LoadStoreLowering
Switching through all the possible memory instruction kinds. Consider
all the required properties that define which intrinsic should be taken
to replace the memory instruction. Though all cases are considered only
non-atomic load/store in global memory with A64 addressing are
supported. The other cases are yet to be supported in further patches.