[Autobackout][FuncReg]Revert of change: ebd9a8f7087f9597318798648d36808a7c9bcf20  Better private memory estimate for indirect calls
Updated the CG analysis algorithm for calculating stack private mem usage.
If a function's address is taken, the function that references that function address now counts as its caller.
This is not always an accurate assumption, since the function pointer can be saved to memory and loaded/called
from a different function group. However, it does provide a better estimate for general cases than not doing this
analysis.
Also, if indirect calls, recursion, or VLA exists, we set the minimum private mem allocation to 8KB.
Also, take into account the IGC flag ForcePerThreadPrivateMemorySize, to change private mem size for debugging.