Fixed a structurizer bug
When a break is recognized, structurizer tried to re-construct if-endif sequence. If
those if-endif can be succesfully re-constructed, the new level hammock is created.

However, the parent field of those children of this new hammock are not set, which
cause wrong control-flow edges. This change fixes it by setting parent fields.