Internalize functions earlier in VC pipeline
Function internalization logic is moved from CMABIAnalysis into a
separate pass GenXLinkageCorruption. This is done because some frontends
are still depend on this behavior and it prevents from making changes to
passes that go before CMABI, e.g. CMImpParam. Also with this patch IR is
no longer changed in an analysis. This pass must be removed once
external functions are fully supported.