Refactor builtins operating on images and samplers
This is a recommit of the following two changes that have been reverted due to regression:
- f2e3bcfe8ea9aa30f60e515f837c0548eca5ec0d "Refactor SPV_INTEL_device_side_avc_motion_estimation builtins"
- 83cf0dbd23457266038b1da6129e14f31d2f002b "Support KHR SPIRV Friendly IR for SPV_INTEL_media_block_io"

This is a preparation for SPIRV Translator switchover to vanilla version.
Vanilla translator produces slightly different form of SPIRV builtins
which is described here:
https://github.com/KhronosGroup/SPIRV-LLVM-Translator/blob/master/docs/SPIRVRepresentationInLLVM.rst

The plan is to support both versions until translator switchover is
done. Once vanilla translator is enabled, old version of SPIRV builtins
will get cleaned up.