Resolve entry point wrapper in SPIRVReader
SPIR-V spec states:
"It is invalid for any function to be targeted by both an OpEntryPoint instruction
and an OpFunctionCall instruction."

In order to satisfy SPIR-V validation tools, there is a plan to overcome
the issue by introducing entrypoint wrapper around functions at SPIRV-LLVM
Translator level. Here is an implementation:
https://github.com/KhronosGroup/SPIRV-LLVM-Translator/pull/1149

IGC SPIRVReader also needs to be able to resolve the wrappers to preserve an
original kernel name. This change is just a partial port from above PR.