[Autobackout][FuncReg]Revert of change: a012be4dccefcacef9f1dc5acd1afdc23790d161  Dont increase spill cost of all address taken variables as this causes RA to spill other short live-ranges. With this change, address taken variables are treated at par with other allocation candidates unless.
The only exception is for all address
takens pointed to by a same address register. If an address register
points to multiple general variables, we need to load all possible
targets to GRF which can lead to a lot of spill/fill code.

Dont increase spill cost of all address taken
variables as this causes RA to spill other short live-ranges. With this
change, address taken variables are treated at par with other
allocation candidates unless.

The only exception is for all address
takens pointed to by a same address register. If an address register
points to multiple general variables, we need to load all possible
targets to GRF which can lead to a lot of spill/fill code.