Implement new algo for NoMask WA
The existing noMask WA has a fundamental issue, which is based on virtual
GRF variables. The correct one should be based on GRF. This change provides
a new approach to apply WA based GRFs.

This change has:
  1. both preRA and postRA WA
  2. new preRA WA re-uses the most of the old preRA.
  3. new postRA WA is a new one. It does:
     3.1 If a BB has a preRA WA, will reuse that preRA WA;
     3.2 Try to optimize as much as possible to avoid redundant mov insts.
         This is done by scanning flag usages for each instructions to see
         if a flag, used in one wa inst, can be re-used for the next wa inst.
  4. The new one will apply to all NoMask insts (the old one only applies to
     the global one (preRA) and spill send/mov (postRA).

With the new optimization added in postRA, hopefully, the performance impact
can be kept minimal.

For now, it is off by default, and will be turned on for VC first.