Generic addrspace support in printf resolution
Format or "%s" stings may be passed into printf through generic address
space pointer. In this case printf resolution will try to resolve the
pointer address space in compile time and call the implementation with
the obtained non-abstract pointer (constant or global). When the address
space cannot be resolved at compile time, an error is returned.