Intrinsics cannot be called indirectly
SPIRV translator does not support unnamed aggregate types, so it adds
some names when translating them. On the other hand, imported emulation
functions do not have such a transformation. That leads to unexpected
behavior when we are linking compiled module with emulation functions
as LLVM::ConstantExpr is created if types differ. Example:
  %cmcl.builtin.i.i.i = tail call { <16 x i32>, <16 x i32> } bitcast
    (%structtype (<16 x i32>, <16 x i32>)* @llvm.genx.addc.v16i32.v16i32
    to { <16 x i32>, <16 x i32> } (<16 x i32>, <16 x i32>)*)
    (<16 x i32> %neg.i131.i, <16 x i32> zeroinit)
  where %structtype = type { <16 x i32>, <16 x i32> }