Remove redundant Funcs from live range
Funcs member in live range in supposed to keep kernels/stackcalls that
are parents to values in this live range. Live ranges are built for
each function group separately. With this, it is impossible to have
more than one function in Funcs set for any live range. Zero functions
are possible for uniform return value when it is the only value in
live range -- uniform ret is recorded in live ranges but it is not
inserted in any function.

Regalloc used Funcs to insert allocated register into regmap by
iterating over Funcs and inserting only if function is current FG
head. With all above, it will always insert only one register for
function group head with exception for unified ret. For unified ret
there will be no entry in regmap (though register will be created).

Remove Funcs to drop redundancy and just insert registers into regmap
for current FG head. There is no functional change for main
compilation flow. Exception is that regmap will now contain entry for
unified ret that will be unused anyway. This also makes regalloc dump
print register for unified ret. Add assertion in dump that register is
created.