VC should produce debuggable kernels by default
This commit enables compilation of kernels that are suitable
for HW debug by default. BTI index zero is considered
reserved (for SIP surface access by the debugger).
Note, that to do a proper debug session one still has to
either inject breakpoint bits to the first instruction
of a kernel (This is enabled by "-g" currently) or to somehow
postpone kernel execution by other means (by passing extra
parameters to runtime or DCD).