Minor change to EU fusion call WA
Make sure to use the same variable for Small EU's call target in call WA.
As those instruction will be marked with M16. They need to be save/restore
as M16 as well.  Using the same var helps call save/restore to only save
a single GRF in which this var is allocated.