EBinary: Do not merge payload arguments those are not contiguous in payload layout
Merge-able payload arguments of by-value-struct aggregate elements must be contiguous
in both host data layout and payload layout.

It's possible that an element is contiguous in host data but is not in the payload.
The case is seen in nested struct argument.
Remove the assert to allow the case, and do not merge the argument.