Support local ID in external, indirect functions
Local IDs are loaded from local_id_table for fixed signature functions
(external and indirect). Table pointer is loaded from implicit args
buffer, group thread ID is used as the offset in the table.

The implemented approach is only valid for platforms that have payload
in memory. The other platforms should be covered in a separate patch.