Refactor memfence_optnone code
EmitMemfence code assumes all parameters of memfence intrinsic are known compile time. This might not be the case in optnone scenarios. To handle this, combinations of "if/else" blocks are generated in the OCL-C built-ins.
This change refactors the code, so that adding additional memfence arguments is easier to maintain without code duplication. Additionally, emmitter assumes worst case scenario, if the constant is not detected nevertheless.