Fix ocloc crash comes from VectorPreProcess pass.
The crash comes from VectorPreProcess pass that splits a large vector
load/store into series of loads/stores of smaller vectors satisfying
certain conditions. For each store, the stored value is taken and split
into smaller vectors based on splitSize parameter and then this partition
is cached in a map "vector value" -> "array of subvectors". The problem
is that there may be a situation where two instructions contain the same
value but its partition is different in each case. The latter instruction
takes an incorrect partition form the map instead of creating new one
and if the incorrect partition contains less subvectors than expected
then index out of bounds error may occur.

The change make the mentioned map more complex i.e. array of subvectors
now depends not only on a vector value but also on on splitSize.