vISA LocalScheduler: add 6-instructions-latency to WAR dependency to a send (2nd try)
Add WAR_SEND_LATENCY=12 for instructions having WAR to sends to hide the latency in between.

A send does not read its srcs quickly. The send message sits in MAO until it grants the use
of the SF. Is until that moment where it starts reading its source. Moreover, if the message
is long (messages can red in average 6 registers), freeing the $x.src dependency can take
several cycles. If there are several EUs competing for the use of the MAO, this time can
elongate further.