Make IGCConstantFolder compatible with LLVM 10 & 11
Previously, `IGCConstantFolder` was switched to contain an
instance of `llvm::ConstantFolder` and proxy common methods through that
so as to prepare for LLVM 11's inheritance scheme changes. However, to
maintain the usability of `IGCConstantFolder` within `llvm::IRBuilder` child classes,
this folder class still has to be part of the `llvm::IRBuilderFolder` hierarchy. The
current changes ensure that, and also achieve build compatibility accross LLVM versions
by abstracting this into a `WrapperLLVM` API.